math_baselineq-main/requirements.txt
accelerate==1.10.0
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1753804962438/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
av==15.0.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1749229842835/work
certifi==2025.8.3
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560556141/work
charset-normalizer==3.4.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1746740462479/work
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1706434688412/work
einops==0.8.1
filelock==3.18.0
frozenlist @ file:///home/conda/feedstock_root/build_artifacts/frozenlist_1752167142843/work
fsspec==2025.7.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1745509389093/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hf-xet==1.1.7
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
huggingface-hub==0.34.3
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna==3.10
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1751310558090/work
multiprocess @ file:///home/conda/feedstock_root/build_artifacts/multiprocess_1724954614853/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1749078300/work
numpy==2.3.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
optree @ file:///home/conda/feedstock_root/build_artifacts/optree_1753455044557/work
packaging==25.0
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1752081722123/work
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751481983695/work
propcache @ file:///home/conda/feedstock_root/build_artifacts/propcache_1744524934837/work
psutil==7.0.0
pyarrow==21.0.0
pybind11 @ file:///D:/bld/bld/rattler-build_pybind11_1752769966/work
pybind11-global @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pybind11-global_1752769938/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2025.7.34
requests==2.32.4
safetensors==0.6.1
sentencepiece @ file:///home/conda/feedstock_root/build_artifacts/sentencepiece-split_1751790858326/work/python
setuptools==80.9.0
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sympy==1.13.1
timm==1.0.19
tokenizers==0.20.3
torch==2.6.0
torchvision==0.21.0
torchvision-extra-decoders @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_torchvision-extra-decoders_1748570842/work
tqdm==4.67.1
transformers==4.46.3
triton==3.2.0
typing_extensions==4.14.1
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3==2.5.0
xxhash @ file:///home/conda/feedstock_root/build_artifacts/python-xxhash_1740594782459/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1749554822108/work
zstandard==0.23.0
